switched enums into enum classes

+++ 2D top down/Weapon.cpp
@@ -27,9 +27,9 @@ void Weapons::AddWeapon(Weapon* weapon)
 		debug.Log("Weapon", "AddWeapon", "RecoilControlRate: " + std::to_string(weapon->recoilControlRate));
 		debug.Log("Weapon", "AddWeapon", "BulletsPerShot: " + std::to_string(weapon->bulletsPerShot));
 		debug.Log("Weapon", "AddWeapon", "ProjectileDistance: " + std::to_string(weapon->projectileDistance));
-		debug.Log("Weapon", "AddWeapon", "BulletType: " + std::to_string(weapon->bulletType));
-		debug.Log("Weapon", "AddWeapon", "fireType: " + std::to_string(weapon->fireType));
-		debug.Log("Weapon", "AddWeapon", "reloadType: " + std::to_string(weapon->reloadType));
+		debug.Log("Weapon", "AddWeapon", "BulletType: " + std::to_string(static_cast<int> (weapon->bulletType)));
+		debug.Log("Weapon", "AddWeapon", "fireType: " + std::to_string(static_cast<int> (weapon->fireType)));
+		debug.Log("Weapon", "AddWeapon", "reloadType: " + std::to_string(static_cast<int> (weapon->reloadType)));
 	}
 }
